big-query-csv-loader/requirements.txt
google-cloud-bigquery
google-cloud-storage
smart-open
pydantic-settings
